+++ lib/quds_interactive.dart
@@ -4,5 +4,5 @@ export 'quds_translation/internals.dart';
 export 'quds_theme/internals.dart';
 export 'quds_interactive/quds_interactive_app.dart';
 
-export 'package:quds_provider/quds_provider.dart';
-export 'package:quds_ui_kit/quds_ui_kit.dart';
+// export 'package:quds_provider/quds_provider.dart';
+// export 'package:quds_ui_kit/quds_ui_kit.dart';