+++ CHANGELOG.md
@@ -22,3 +22,22 @@
 ## 0.0.2+1
 * Re-wrote README.md content
 
+## 0.0.3+1
+* Add some documentations.
+* Add support of custom fonts.
+
+## 0.0.3+3
+* Added `customBuilder` for `QudsInteractiveBuilder`.
+
+## 0.0.3+7
+* Updated dependencies.
+
+## 0.0.4+2
+* Added `QudsSelectThemeListTile` + `QudsSelectLanguageListTile`.
+* Added `QudsSelectThemeIconButton` + `QudsSelectLanguageIconButton`.
+
+## 0.0.4+4
+* Updated dependencies
+
+## 0.0.4+5
+* Added support of titles for `QudsInteractiveApp` => `title` + `generateTitle`